// src/app/api/keymap.ts
export enum StarCitizenKey {
  COMMS = 'COMMS',
  MOBI_GLAS = 'MOBI_GLAS',
  STAR_MAP = 'STAR_MAP',
  DECOUPLE = 'DECOUPLE',
  GSAFE = 'GSAFE',
  AFTER_BURNER = 'AFTER_BURNER',
  MAX_THRUST = 'MAX_THRUST',
  ESP = 'ESP',
  SCM_LIMITER = 'SCM_LIMITER',
  FREE_LOOK = 'FREE_LOOK',
  SPOOL_DRIVE = 'SPOOL_DRIVE',
  QUANTUM_JUMP = 'QUANTUM_JUMP',
  SELF_DESTRUCT = 'SELF_DESTRUCT',
  EJECT = 'EJECT',
  LOCK = 'LOCK',
  UNLOCK = 'UNLOCK',
  SCANNING = 'SCANNING',
  LIGHTS = 'LIGHTS',
  OPEN_DOORS = 'OPEN_DOORS',
  CLOSE_DOORS = 'CLOSE_DOORS',
}